requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
scikit-learn>=1.0.0
joblib>=1.0.0
googletrans==4.0.0-rc1
pdfminer.six>=20201018
python-docx>=0.8.11
requests>=2.28.0